Functions

Support Functions

TransComp.get_input_data — Function

get_input_data(path_to_source_file::String)

This function reads the input data and checks requirements for the content of the file.

Arguments

pathtosource_file::String: path to the source file

Returns

data_dict::Dict: dictionary with the input data

TransComp.parse_data — Function

parse_data(data_dict::Dict)

Parses the input data into the corresponding parameters in struct format from structs.jl.

Arguments

data_dict::Dict: dictionary with the input data

Returns

data_structures::Dict: dictionary with the parsed data

TransComp.create_model — Function

create_model(data_structures, case_name::String, optimizer)

Definition of JuMP.model and adding of variables.

Arguments

data_structures: dictionary with the input data and parsing of the input parameters

case_name::String: name of the case

optimizer: optimizer to use for the model

Returns

model::JuMP.Model: JuMP model with the variables added

data_structures::Dict: dictionary with the input data

TransComp.save_results — Function

save_results(model::Model, case::String, data_structures::Dict, write_to_file::Bool = true, folder_for_results::String = "results")

Saves the results of the optimization model to YAML files.

Arguments

model::Model: JuMP model

case::String: name of the case

data_structures::Dict: dictionary with the input data

write_to_file::Bool: whether to write results to file (default: true)

folder_for_results::String: name of the folder to save the results (default: "results")