Transcomp Model Documentation

The Transcomp model is a linear optimization model developed within the EU Horizon project iDesignRES in workpage 1. It is implemented in Julia and is based on the JuMP package.

The Transcomp model is aimed at modeling developments in subsectors of the transport sector. It is designed in a way that it can be adoptable in its complexity in the representation of the transport sector - being flexible in its ability to adjust in to the level of availibility of granular data - and being costumizable in the scope if representation of the the transport sector, i.e. included transport segments. Further, features are provided that allow the soft linking to traditional transport models and energy system models.

Manual outline

Main.TransComp.base_define_variables — Method

base_define_variables(model::Model, data_structures::Dict)

Defines the variables for the model.

Arguments

model::Model: JuMP model

data_structures::Dict: dictionary with the input data

Main.TransComp.check_input_file — Method

check_input_file(path_to_source_file::String)

Check if the input file exists and is a YAML file.

Arguments

path_to_source_file::String: The path to the input file.

Main.TransComp.check_model_parametrization — Method

check_model_parametrization(data_dict::Dict, required_keys::Vector{String})

Check if the required keys are present in the model data.

Arguments

data_dict::Dict: The input data.

Returns

Bool: True if the required keys are present, false otherwise.

Main.TransComp.check_required_keys — Method

check_required_keys(data_dict::Dict, required_keys::Vector{String})

Check if the required keys are present in the input data.

Arguments

data_dict::Dict: The input data.

Main.TransComp.constraint_demand_coverage — Method

constraint_demand_coverage(model::JuMP.Model, data_structures::Dict)

Creates constraint for demand coverage.

Arguments

model::JuMP.Model: JuMP model

data_structures::Dict: dictionary with the input data

Main.TransComp.constraint_fueling_demand — Method

constraint_fueling_demand(model::JuMP.Model, data_structures::Dict)

Constraints for fueling demand at nodes and edges.

Arguments

model::JuMP.Model: JuMP model

data_structures::Dict: dictionary with the input data

Main.TransComp.constraint_fueling_infrastructure — Method

constraint_vehicle_purchase(model::JuMP.Model, data_structures::Dict)

Constraints for the sizing of fueling infrastructure at nodes and edges.

Arguments

model::JuMP.Model: JuMP model

data_structures::Dict: dictionary with the input data

Main.TransComp.constraint_mode_shift — Method

constraint_mode_shift(model::JuMP.Model, data_structures::Dict)

Constraints for the rate of the mode shfit.

Arguments

model::JuMP.Model: JuMP model

data_structures::Dict: dictionary with the input data

Main.TransComp.constraint_monetary_budget — Method

constraint_vehicle_purchase(model::JuMP.Model, data_structures::Dict)

Creates constraints for monetary budget for vehicle purchase by route.

Arguments

model::JuMP.Model: JuMP model

data_structures::Dict: dictionary with the input data

Main.TransComp.constraint_vehicle_aging — Method

constraint_vehicle_aging(model::JuMP.Model, data_structures::Dict)

Creates constraints for vehicle aging.

Arguments

model::JuMP.Model: JuMP model

data_structures::Dict: dictionary with the input data

Returns

model::JuMP.Model: JuMP model with the constraints added

Main.TransComp.constraint_vehicle_sizing — Method

constraint_vehicle_sizing(model::JuMP.Model, data_structures::Dict)

Creates constraint for vehicle sizing.

Arguments

model::JuMP.Model: JuMP model

data_structures::Dict: dictionary with the input data

Main.TransComp.constraint_vehicle_stock_shift — Method

constraint_vehicle_stock_shift(model::JuMP.Model, data_structures::Dict)

Constraints for vehicle stock turnover.

Arguments

model::JuMP.Model: JuMP model

data_structures::Dict: dictionary with the input data

Main.TransComp.create_m_tv_pairs — Method

create_m_tv_pairs(techvehicle_list::Vector{TechVehicle}, mode_list::Vector{Mode})

Creates a set of pairs of mode and techvehicle IDs.

Arguments

techvehicle_list::Vector{TechVehicle}: list of techvehicles

mode_list::Vector{Mode}: list of modes

Returns

mtvpairs::Set: set of pairs of mode and techvehicle IDs

Main.TransComp.create_model — Method

create_model(model::JuMP.Model, data_structures::Dict)

Definition of JuMP.model and adding of variables.

Arguments

model::JuMP.Model: JuMP model

data_structures::Dict: dictionary with the input data and parsing of the input parameters

Returns

model::JuMP.Model: JuMP model with the variables added

data_structures::Dict: dictionary with the input data

Main.TransComp.create_p_r_k_e_set — Method

create_p_r_k_e_set(odpairs::Vector{Odpair})

Creates a set of pairs of product, odpair, path, and element IDs.

Arguments

odpairs::Vector{Odpair}: list of odpairs

Returns

prkepairs::Set: set of pairs of product, odpair, path, and element IDs

Main.TransComp.create_p_r_k_n_set — Method

create_p_r_k_n_set(odpairs::Vector{Odpair})

Creates a set of pairs of product, odpair, path, and element IDs.

Arguments

odpairs::Vector{Odpair}: list of odpairs

Returns

prknpairs::Set: set of pairs of product, odpair, path, and element IDs

Main.TransComp.create_p_r_k_set — Method

create_p_r_k_set(odpairs::Vector{Odpair})

Creates a set of pairs of product, odpair, and path IDs.

Arguments

odpairs::Vector{Odpair}: list of odpairs

Returns

prk_pairs::Set: set of pairs of product, odpair, and path IDs

Main.TransComp.create_r_k_set — Method

create_r_k_set(odpairs::Vector{Odpair})

Creates a set of pairs of odpair and path IDs.

Arguments

odpairs::Vector{Odpair}: list of odpairs

Returns

rkpairs::Set: set of pairs of odpair and path IDs

Main.TransComp.create_tv_id_set — Method

create_tv_id_set(techvehicle_list::Vector{TechVehicle}, mode_list::Vector{Mode})

Creates a list of techvehicle IDs.

Arguments

techvehicle_list::Vector{TechVehicle}: list of techvehicles

mode_list::Vector{Mode}: list of modes

Returns

techvehicleids2::Set: set of techvehicle IDs

Main.TransComp.create_v_t_set — Method

create_v_t_set(techvehicle_list::Vector{TechVehicle})

Creates a set of pairs of techvehicle IDs.

Arguments

techvehicle_list::Vector{TechVehicle}: list of techvehicles

Returns

tvpairs::Set: set of pairs of techvehicle IDs

Main.TransComp.depreciation_factor — Method

depreciation_factor(y, g)

Calculate the depreciation factor for a vehicle based on its age.

Arguments

y::Int: The year of the vehicle.

g::Int: The year the vehicle was purchased.

Returns

Float64: The depreciation factor.

Main.TransComp.get_input_data — Method

get_input_data(path_to_source_file::String)

This function reads the input data and checks requirements for the content of the file.

Arguments

pathtosource_file::String: path to the source file

Returns

data_dict::Dict: dictionary with the input data

Main.TransComp.objective — Method

objective(model::Model, data_structures::Dict)

Definition of the objective function for the optimization model.

Arguments

model::Model: JuMP model

data_structures::Dict: dictionary with the input data

Main.TransComp.parse_data — Method

parse_data(data_dict::Dict)

Parses the input data into the corresponding parameters in struct format from structs.jl.

Arguments

data_dict::Dict: dictionary with the input data

Returns

data_structures::Dict: dictionary with the parsed data

Main.TransComp.save_results — Method

save_results(model::Model, case_name::String)

Saves the results of the optimization model to YAML files.

Arguments

model::Model: JuMP model

case_name::String: name of the case

Returns

output_file::String: name of the output file

Main.TransComp.Edge — Type

Edge

An 'Edge' represents a connection between two nodes and is a representation of connecting transport infrastructure.

Fields

id::Int: unique identifier of the edge

name::String: name of the connection

length::Float64: length of the connection in km

from::Node: the node from which the edge starts

to::Node: the node to which the edge ends

Main.TransComp.Emission_constraints_by_mode — Type

Emission_constraints_by_mode

An 'Emissionconstraintsby_mode' describes emissions constrained for a mode.

Fields

id::Int: unique identifier of the emission constraint

mode::Mode: mode of transport

emission::Float64: emission constraint of the vehicle type

year::Int: year of the expected emission constraint

Main.TransComp.Emission_constraints_by_year — Type

Emission_constraints_by_year

An 'Emissionconstraintsby_year' describes an emission goal for a specific year for the total emissions.

Fields

id::Int: unique identifier of the emission constraint

emission::Float64: emission constraint

year::Int: year of the expected emission constraint

Main.TransComp.F_init_mode_share — Type

F_init_mode_share

A 'Finitmode_share' describes the initial mode share of a transport mode and is an attribute of a travel connection.

Fields

id::Int: unique identifier of the initial mode share

mode::Mode: mode of transport

share::Float64: share of the mode in the initial mode share

f_init::Float64: initial number of trips in p/year or t/year

Main.TransComp.FinancialStatus — Type

FinancialStatus

A 'FinancialStatus' describes a demographic group based on what there average budget for transportation-related expenses is.

Fields

id::Int: unique identifier of the financial status

name::String: name of the financial status

VoT: value of time in €/h

monetary_budget_operational: budget for operational costs in €/year

monetary_budget_operational_lb: lower bound of the budget for operational costs in €/year

monetary_budget_operational_ub: upper bound of the budget for operational costs in €/year

monetary_budget_purchase: budget for purchasing costs in €/year

monetary_budget_purchase_lb: lower bound of the budget for purchasing costs in €/year

monetary_budget_purchase_ub: upper bound of the budget for purchasing costs in €/year

Main.TransComp.Fuel — Type

Fuel

A 'Fuel' represents the energy source used for the vehicle propulsion.

Fields

id::Int: unique identifier of the fuel

name::String: name of the fuel

emission_factor::Float64: emission factor of the fuel in gCO2/kWh

cost_per_kWh: cost per kWh of the fuel in €

cost_per_kW: cost per kW of the fuel in €

Main.TransComp.Generation — Type

Generation

A 'Generation' referes to model year of a vehicle type to include the variation in vehicle performance parameters depending on when a vehicle was produced.

Fields

id::Int: unique identifier of the generation

year::Int: year of the generation

y::Int: year of the generation relative to the initial year of the optimization horizon

Main.TransComp.InitialVehicleStock — Type

InitialVehicleStock

An 'InitialVehicleStock' represents a vehicle fleet that exisits at the initial year of the optimization horizon.

Fields

id::Int: unique identifier of the initial vehicle stock

techvehicle::TechVehicle: vehicle type and technology of the vehicle

year_of_purchase::Int: year in which the vehicle was purchased

stock::Float64: number of vehicles of this type in the initial vehicle stock

Main.TransComp.Mode — Type

Mode

A 'Mode' represents a transport mode. Transport modes may differ either by the infrastructure used (for example, road vs. rail) or by the used vehicle type (for example, private passenger car vs. bus) that directly influences the travel time but excludes a differentiation based on technology.

Fields

id::Int: unique identifier of the mode

name::String: name of the mode

quantify_by_vehs::Bool: if for this mode vehicles stock is sized or not. If this mode is considered with levelized costs, including the costs for vehicles and related costs.

cost_per_ukm::Array{Float64, 1}: cost per km in €/km (only relevant when quantifybyvehs is false)

emission_factor::Array{Float64,1}: emission factor of the mode in gCO2/ukm (only relevant when quantifybyvehs is false)

Main.TransComp.Mode_share_max — Type

Mode_share_max

Maximum mode shares of a transport mode independent of year, i.e. over total horizon.

Fields

id::Int: unique identifier of the mode share

mode::Mode: mode of transport

share::Float64: maximum share of the mode

financial_status::Array{FinancialStatus, 1}: array of financial status that is affected by this mode share constraint

Main.TransComp.Mode_share_max_by_year — Type

Mode_share_max_by_year

Maximum mode shares of a transport mode in a specific year.

Fields

id::Int: unique identifier of the mode share

mode::Mode: mode of transport

share::Float64: maximum share of the mode

year::Int: year of the maximum mode share

financial_status::Array{FinancialStatus, 1}: financial status that is affected by this mode share constraint

Main.TransComp.Mode_share_min — Type

Mode_share_min

Maximum mode shares of a transport mode independent of year, i.e. over total horizon.

Fields

id::Int: unique identifier of the mode share

mode::Mode: mode of transport

share::Float64: maximum share of the mode

financial_status::Array{FinancialStatus, 1}: array of financial status that is affected by this mode share constraint

Main.TransComp.Mode_share_min_by_year — Type

Mode_share_min_by_year

Minimum mode shares of a transport mode in a specific year.

Fields

id::Int: unique identifier of the mode share

mode::Mode: mode of transport

share::Float64: minimum share of the mode

year::Int: year of the minimum mode share

financial_status::Array{FinancialStatus, 1}: financial status that is affected by this mode share constraint

Main.TransComp.Node — Type

Node

A 'Node' represents geographic region.

Fields

id::Int: unique identifier of the node

name::String: name the region

Main.TransComp.Odpair — Type

Odpair

An 'Odpair' describes transport demand. It may take place between two regions but origin and destination may al so

Fields

id::Int: unique identifier of the odpair

origin::Node: origin of the transport demand

destination::Node: destination of the transport demand

paths::Array{Path, 1}: possible paths between origin and destination

F: number of trips in p/year or t/year

Main.TransComp.Path — Type

Path

A 'Path' represents a possible route between two nodes. This sequence includes the nodes that are passed through and the length of the path.

Fields

id::Int: unique identifier of the path

name::String: name of the path

length::Float64: length of the path in km

sequence: sequence of nodes and edges that are passed through

Main.TransComp.Product — Type

Product

A 'Product' represents either a good or a service that is being transported. This may include passengers, or different types of products in the freight transport. The differentiation of transported products related to the different needs for transportation and, therefore, different possible sets of transport modes, vehicle types and drivetrain technologies are available for transport.

Fields

id::Int: unique identifier of the product

name::String: name of the product

Main.TransComp.Regiontype — Type

Regiontype

A 'Regiontype' describes a region based on its characteristics that induces differences in transportation needs (for example, urban vs. rural area).

Fields

id::Int: unique identifier of the regiontype

name::String: name of the regiontype

speed::Float64: average speed in km/h

costs_var::Array{Float64, 1}: variable costs in €/vehicle-km

costs_fix::Array{Float64, 1}: fixed costs in €/year

Main.TransComp.TechVehicle — Type

TechVehicle

A 'TechVehicle' represents a vehicle that is used for transportation. This includes the vehicle type, the technology used in the vehicle, the capital and maintenance costs, the load capacity, the specific consumption, the lifetime, the annual range, the number of vehicles of this type, the battery capacity, and the peak charging power.

Main.TransComp.TechVehicle_share_max — Type

TechVehicle_share_max

Maximum vehicle type shares of a TechVehicle independent of year, i.e. over total horizon.

Fields

id::Int: unique identifier of the TechVehicle share

techvehicle::TechVehicle: TechVehicle

share::Float64: maximum share of the TechVehicle

financial_status::Array{FinancialStatus, 1}: array of financial status that is affected by this TechVehicle share constraint

Main.TransComp.TechVehicle_share_max_by_year — Type

TechVehicle_share_max_by_year

Maximum vehicle type shares of a TechVehicle in a specific year.

Fields

id::Int: unique identifier of the TechVehicle share

techvehicle::TechVehicle: TechVehicle

share::Float64: maximum share of the TechVehicle

year::Int: year of the maximum TechVehicle share

financial_status::Array{FinancialStatus, 1}: financial status that is affected by this TechVehicle share constraint

Main.TransComp.TechVehicle_share_min — Type

TechVehicle_share_min

Minimum vehicle type shares of a TechVehicle independent of year, i.e. over total horizon.

Fields

id::Int: unique identifier of the TechVehicle share

techvehicle::TechVehicle: TechVehicle

share::Float64: minimum share of the TechVehicle

financial_status::Array{FinancialStatus, 1}: array of financial status that is affected by this TechVehicle share constraint

Main.TransComp.TechVehicle_share_min_by_year — Type

TechVehicle_share_min_by_year

Minimum vehicle type shares of a TechVehicle in a specific year.

Fields

id::Int: unique identifier of the TechVehicle share

techvehicle::TechVehicle: TechVehicle

share::Float64: minimum share of the TechVehicle

year::Int: year of the minimum TechVehicle share

financial_status::Array{FinancialStatus, 1}: financial status that is affected by this TechVehicle share constraint

Main.TransComp.Technology — Type

Technology

A 'Technology' represents the drivetrain technology used in the vehicle.

Fields

id::Int: unique identifier of the technology

name::String: name of the technology

fuel::Fuel: fuel used by the technology

Main.TransComp.Technology_share_max — Type

Technology_share_max

Maximum technology shares of a vehicle technology independent of year, i.e. over total horizon.

Fields

id::Int: unique identifier of the technology share

technology::Technology: vehicle technology

share::Float64: maximum share of the technology

financial_status::Array{FinancialStatus, 1}: array of financial status that is affected by this technology share constraint

Main.TransComp.Technology_share_max_by_year — Type

Technology_share_max_by_year

Maximum technology shares of a vehicle technology in a specific year.

Fields

id::Int: unique identifier of the technology share

technology::Technology: vehicle technology

share::Float64: maximum share of the technology

year::Int: year of the maximum technology share

financial_status::Array{FinancialStatus, 1}: financial status that is affected by this technology share constraint

Main.TransComp.Technology_share_min — Type

Technology_share_min

Minimum technology shares of a vehicle technology independent of year, i.e. over total horizon.

Fields

id::Int: unique identifier of the technology share

technology::Technology: vehicle technology

share::Float64: minimum share of the technology

financial_status::Array{FinancialStatus, 1}: array of financial status that is affected by this technology share constraint

Main.TransComp.Technology_share_min_by_year — Type

Technology_share_min_by_year

Minimum technology shares of a vehicle technology in a specific year.

Fields

id::Int: unique identifier of the technology share

technology::Technology: vehicle technology

share::Float64: minimum share of the technology

year::Int: year of the minimum technology share

financial_status::Array{FinancialStatus, 1}: financial status that is affected by this technology share constraint

Main.TransComp.Transportation_speeds — Type

Transportation_speeds

A 'Speed' describes the speed of a vehicle type in a specific year.

Fields

Main.TransComp.VehicleType_share_max — Type

VehicleType_share_max

Maximum vehicle type shares of a vehicle type independent of year, i.e. over total horizon.

Fields

id::Int: unique identifier of the vehicle type share

vehicle_type::Vehicletype: vehicle type

share::Float64: maximum share of the vehicle type

financial_status::Array{FinancialStatus, 1}: array of financial status that is affected by this vehicle type share constraint

Main.TransComp.VehicleType_share_max_by_year — Type

VehicleType_share_max_by_year

Maximum vehicle type shares of a vehicle type in a specific year.

Fields

id::Int: unique identifier of the vehicle type share

vehicle_type::Vehicletype: vehicle type

share::Float64: maximum share of the vehicle type

year::Int: year of the maximum vehicle type share

financial_status::Array{FinancialStatus, 1}: financial status that is affected by this vehicle type share constraint

Main.TransComp.VehicleType_share_min — Type

VehicleType_share_min

Minimum vehicle type shares of a vehicle type independent of year, i.e. over total horizon.

Fields

id::Int: unique identifier of the vehicle type share

vehicle_type::Vehicletype: vehicle type

share::Float64: minimum share of the vehicle type

financial_status::Array{FinancialStatus, 1}: array of financial status that is affected by this vehicle type share constraint

Main.TransComp.VehicleType_share_min_by_year — Type

VehicleType_share_min_by_year

Minimum vehicle type shares of a vehicle type in a specific year.

Fields

id::Int: unique identifier of the vehicle type share

vehicle_type::Vehicletype: vehicle type

share::Float64: minimum share of the vehicle type

year::Int: year of the minimum vehicle type share

financial_status::Array{FinancialStatus, 1}: financial status that is affected by this vehicle type share constraint

Main.TransComp.Vehicletype — Type

Vehicletype

A 'Vehicletype' represents a type of vehicle that is used for transportation. This may be for example, small passenger cars, buses, or light-duty trucks.

Fields

id::Int: unique identifier of the vehicle type

name::String: name of the vehicle type

mode::Mode: mode of transport that the vehicle type is used for

product::Product: product that the vehicle type is used for